Added mapBox dependency, tested authentication to MapBox service, HomeScreen moved to separate feature
