+++ feature/home/build.gradle.kts
@@ -12,6 +12,10 @@ dependencies {
     implementation(projects.library.navigation)
     implementation(projects.library.resources)
 
+    implementation(libs.accompanist.permissions)
+
+    implementation(libs.mapbox.android)
+
     implementation(libs.compose.ui)
     implementation(libs.compose.ui.preview)
     implementation(libs.compose.material)
